Add bias_point option
Allows setting the bias point to an arbitrary time, instead of "right now"
when determining past or future.

Also adds a shortcut to the bias option, "future_date" or "past_date", to
set it to midnight today.

When parsing a partial date only, without any care for time, allows for
example, on Oct 21, 2015, "oct 21" to parse as 2015-10-21 all day, instead
of 2016-10-21 (when bias is set to 'future_date').